fix: luckphrases -> luckphrase

+++ src/main/java/com/luckkids/domain/push/Push.java
@@ -13,38 +13,38 @@ import lombok.NoArgsConstructor;
 @Entity
 public class Push extends BaseTimeEntity {
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
 
-	@ManyToOne(fetch = FetchType.LAZY)
-	private User user;
+    @ManyToOne(fetch = FetchType.LAZY)
+    private User user;
 
-	private String pushToken;
+    private String pushToken;
 
-	private String deviceId;
+    private String deviceId;
 
-	@Builder
-	private Push(User user, String pushToken, String deviceId) {
-		this.user = user;
-		this.pushToken = pushToken;
-		this.deviceId = deviceId;
-	}
+    @Builder
+    private Push(User user, String pushToken, String deviceId) {
+        this.user = user;
+        this.pushToken = pushToken;
+        this.deviceId = deviceId;
+    }
 
-	public static Push of(User user, String pushToken, String deviceId){
-		return Push.builder()
-			.user(user)
-			.pushToken(pushToken)
-			.deviceId(deviceId)
-			.build();
-	}
+    public static Push of(User user, String pushToken, String deviceId) {
+        return Push.builder()
+            .user(user)
+            .pushToken(pushToken)
+            .deviceId(deviceId)
+            .build();
+    }
 
-	public void updatePushToken(String pushToken){
-		this.pushToken = pushToken;
-	}
+    public void updatePushToken(String pushToken) {
+        this.pushToken = pushToken;
+    }
 
-	public void setUser(User user){
-		this.user = user;
-		user.getPushes().add(this);
-	}
+    public void setUser(User user) {
+        this.user = user;
+        user.getPushes().add(this);
+    }
 }
