+++ src/test/java/com/luckkids/api/service/withdrawReason/WithdrawReasonServiceTest.java
@@ -5,7 +5,6 @@ import com.luckkids.api.service.withdrawReason.request.WithdrawReasonCreateServi
 import com.luckkids.api.service.withdrawReason.response.WithdrawReasonCreateResponse;
 import com.luckkids.domain.withdrawReason.WithdrawReason;
 import com.luckkids.domain.withdrawReason.WithdrawReasonRepository;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,14 +20,14 @@ public class WithdrawReasonServiceTest extends IntegrationTestSupport {
 
     @Test
     @DisplayName("탈퇴사유를 저장한다.")
-    void withdrawReasonTest(){
+    void withdrawReasonTest() {
         WithdrawReasonCreateServiceRequest withdrawReasonCreateServiceRequest = WithdrawReasonCreateServiceRequest.builder()
             .reason("잘 사용하지 않는 앱이예요.")
             .build();
 
         WithdrawReasonCreateResponse withdrawReasonCreateResponse = withdrawReasonService.create(withdrawReasonCreateServiceRequest);
 
-        WithdrawReason savedWithdrawReason = withdrawReasonRepository.findById(withdrawReasonCreateResponse.getId()).get();
+        WithdrawReason savedWithdrawReason = withdrawReasonRepository.findById((long) withdrawReasonCreateResponse.getId()).get();
 
         assertThat(savedWithdrawReason.getReason()).isEqualTo("잘 사용하지 않는 앱이예요.");
